Auto encoders were used to colorize black and white images.
Autoencoders are a type of neural network architecture that is commonly used for unsupervised learning. The basic idea behind autoencoders is to learn a compact representation (encoding) of the input data and then reconstruct the input data from this representation. In the context of image colorization, autoencoders are particularly useful for learning meaningful representations of grayscale images and generating their colorized versions.